requirements.txt
absl-py==2.1.0
addict==2.4.0
APScheduler==3.11.0
astunparse==1.6.3
basicsr==1.4.2
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
diffusers==0.31.0
facexlib==0.3.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gfpgan==1.3.8
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
Jinja2==3.1.4
keras==3.7.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.9
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
PyYAML==6.0.2
realesrgan==0.3.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
schedule==1.2.2
scikit-image==0.25.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
tb-nightly==2.19.0a20241217
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tifffile==2024.12.12
tokenizers==0.21.0
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yapf==0.43.0
zipp==3.21.0
